slider taken from portfolio page to landing page
