+++ src/components/PortfolioDetail/ProjectDescription.tsx
@@ -6,47 +6,7 @@ import { ProjectsInterface } from '../../screens/interfaces'
 
 const PortfolioDescription = ({ project }: { project: ProjectsInterface }) => {
     
-  const techLogos = [
-    "devicon-django-plain colored",
-    "devicon-react-plain colored",
-    "devicon-javascript-plain colored",
-    "devicon-css3-plain colored",
-    "devicon-tailwindcss-plain colored",
-    "devicon-typescript-plain colored",
-    "devicon-html5-plain colored",
-    "devicon-qt-plain colored",
-    "devicon-bootstrap-plain colored",
-    "devicon-digitalocean-plain colored",
-    "devicon-docker-plain colored",
-    "devicon-electron-original colored",
-    "devicon-nodejs-plain colored",
-    "devicon-npm-original-wordmark colored",
-    "devicon-numpy-plain colored",
-    "devicon-pandas-plain colored",
-    "devicon-oauth-plain colored",
-    "devicon-mysql-plain colored",
-    "devicon-postgresql-plain colored",
-    "devicon-rabbitmq-plain colored",
-    "devicon-raspberrypi-plain colored",
-    "devicon-redis-plain colored",
-    "devicon-selenium-plain colored",
-    "devicon-slack-plain colored",
-    "devicon-sqlite-plain colored",
-    "devicon-pytest-plain colored",
-    "devicon-pypi-plain colored",
-    "devicon-pycharm-plain colored",
-    "devicon-vscode-plain colored",
-    "devicon-nginx-plain colored",
-    "devicon-nextjs-plain colored",
-    "devicon-mongodb-plain colored",
-    "devicon-jquery-plain colored",
-    "devicon-json-plain colored",
-    "devicon-github-plain colored",
-    "devicon-git-plain colored",
-    "devicon-figma-plain colored",
-    "devicon-django-plain colored",
-    "devicon-python-plain colored",
-  ];
+
 
   return (
     <section id="description" className="mt-20">
@@ -69,51 +29,7 @@ const PortfolioDescription = ({ project }: { project: ProjectsInterface }) => {
           <p className="description-text">{project?.description}</p>
         </div>
       </div>
-      <div className="slider-section">
-        <div
-          className="slider"
-          data-reverse="true"
-          style={
-            {
-              "--width": "130px",
-              "--height": "50px",
-              "--quantity": 39,
-            } as React.CSSProperties
-          }
-        >
-          <div className="list">
-            {techLogos.map((logo, index) => (
-              <div key={index}>
-                <div
-                  className="item"
-                  style={
-                    {
-                      "--position": index,
-                      "font-size": "25px",
-                    } as React.CSSProperties
-                  }
-                >
-                  {logo.startsWith("https") ? (
-                    <img
-                      style={
-                        {
-                          "--width": "25px",
-                          "--height": "25px",
-                          "--quantity": 39,
-                        } as React.CSSProperties
-                      }
-                      src={logo}
-                      alt=""
-                    />
-                  ) : (
-                    <i className={logo}></i>
-                  )}{" "}
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </div>
+
     </section>
   );
 };
